Fix wrong error message when reaching "maxProjects" quota
When reporting the reason of a project import failure, the GitCloneStep
always assumed it was because of a repository with the same name already
exists.

This however, was wrong, since the failure could have been caused by any
number of reasons, including for example, validation errors due to quota
enforcement.

Rather than treating ResourceConflictException as "git destination
already exists", we now expose the actual error message that caused the
failure.

For example, in case of a quota validation error, the reported message
now is something similar to:

``` Unable to create repository syntonyze/playground:Project cannot be
created because a quota for the namespace 'syntonyze/** allows at most 3
projects and 3 projects already exist.
```

Bug: Issue 445765274
Change-Id: I7ab32c23140288b3c259d4b8e5fcb3135f14ce7c